split out functionality related to SQL columns into its own module
